// apps/web/src/lib/api/generated/student-feedback/student-feedback.zod.ts
export const StudentFeedbackQuestionIndexResponse = zod.object({
  "data": zod.array(zod.object({
  "id": zod.string(),
  "institution_id": zod.string(),
  "question": zod.string(),
  "order_index": zod.number(),
  "created_at": zod.iso.datetime({}).nullable(),
  "updated_at": zod.iso.datetime({}).nullable()
}))
})